Add files via upload
Change Log:

Flagged files now are actually copied to the Flagged folder.
If the user chooses NOT to browse the flagged files automatically, program DOES NOT copy the flagged files into the lvl2 directory and DOES NOT create the corresponding entries in the extract (Log) txt file. The Po edge is removed from these files.
If the user chooses to browse the files manually, the old routine is used: new Po edge is set by the user and the flagged files are evaluated together with the other.